Edits, styling and texts

+++ app/Http/Controllers/UsersController.php
@@ -7,6 +7,24 @@ use App\Models\User;
 
 class UsersController extends Controller
 {
+    public function download()
+    {
+        $users = User::all();
+        $filename = "users.csv";
+        $delimiter=",";
+        header('Content-Type: application/csv');
+        header('Content-Disposition: attachment; filename="'.$filename.'";');
+        $f = fopen('php://output', 'w');
+        fputcsv($f, array('name','email','organization','category','platform_role'), $delimiter);
+        foreach ($users as $user) {
+            $role = "user";
+            if($user->is_admin) { $role = "admin"; }
+            else if($user->is_editor) { $role = "editor"; }
+            $line = array($user->name,$user->email,$user->org_name,$user->org_category,$role);
+            fputcsv($f, $line, $delimiter);
+        }
+    }
+
     public function manage()
     {
         $users = User::all();
@@ -27,7 +45,7 @@ class UsersController extends Controller
 
     public function delete(Request $request, User $user)
     {
-        if(isset($_POST['delete'])) {
+        if(isset($_POST['delete']) && $user->id !== 1) {
     		$user->delete();
     	}
         $users = User::all();
